requirements.txt
streamlit==1.28.1
pandas==2.0.3
plotly==5.17.0
numpy==1.24.3
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
scikit-learn==1.3.0
joblib==1.3.2
python-dateutil==2.8.2
pytz==2023.3